requirements.txt
gitdb==4.0.11
GitPython==3.1.40
python-dotenv==1.0.0
smmap==5.0.1
watchdog==3.0.0
